feat(schema): change numeric type

+++ libs/rs/shared/src/database/models/forza.rs
@@ -1,6 +1,5 @@
 #[cfg(feature = "db")]
 use crate::database::schema::forza_data;
-use bigdecimal::BigDecimal;
 use bincode::{Decode, Encode};
 use binrw::BinRead;
 use chrono::NaiveDateTime;
@@ -27,12 +26,10 @@ use strum_macros::{AsRefStr, Display, EnumIter};
     Decode,
     EnumIter,
     AsRefStr,
-    // DbEnum,
 )]
 #[br(little, repr(i32))]
 #[cfg_attr(feature = "db", derive(DbEnum))]
 #[cfg_attr(feature = "db", DbValueStyle = "verbatim")]
-// #[DbValueStyle = "verbatim"]
 pub enum ForzaType {
     #[default]
     FH5,
@@ -48,9 +45,6 @@ pub enum ForzaType {
 )]
 #[cfg_attr(feature = "db", diesel(table_name = forza_data))]
 #[cfg_attr(feature = "db", diesel(check_for_backend(diesel::pg::Pg)))]
-// #[derive(Debug, Clone, PartialEq, Insertable, Selectable, Queryable, Identifiable)]
-// #[diesel(table_name = forza_data)]
-// #[diesel(check_for_backend(diesel::pg::Pg))]
 pub struct ForzaData {
     pub id: String,
     pub game_type: ForzaType,
@@ -64,92 +58,92 @@ pub struct ForzaData {
     pub num_cylinders: i32,
     pub track_id: Option<i32>,
 
-    pub engine_max_rpm: BigDecimal,
-    pub engine_idle_rpm: BigDecimal,
-    pub engine_current_rpm: BigDecimal,
+    pub engine_max_rpm: f32,
+    pub engine_idle_rpm: f32,
+    pub engine_current_rpm: f32,
 
-    pub acceleration_x: BigDecimal,
-    pub acceleration_y: BigDecimal,
-    pub acceleration_z: BigDecimal,
+    pub acceleration_x: f32,
+    pub acceleration_y: f32,
+    pub acceleration_z: f32,
 
-    pub velocity_x: BigDecimal,
-    pub velocity_y: BigDecimal,
-    pub velocity_z: BigDecimal,
+    pub velocity_x: f32,
+    pub velocity_y: f32,
+    pub velocity_z: f32,
 
-    pub angular_velocity_x: BigDecimal,
-    pub angular_velocity_y: BigDecimal,
-    pub angular_velocity_z: BigDecimal,
+    pub angular_velocity_x: f32,
+    pub angular_velocity_y: f32,
+    pub angular_velocity_z: f32,
 
-    pub yaw: BigDecimal,
-    pub pitch: BigDecimal,
-    pub roll: BigDecimal,
+    pub yaw: f32,
+    pub pitch: f32,
+    pub roll: f32,
 
-    pub normalized_suspension_travel_front_left: BigDecimal,
-    pub normalized_suspension_travel_front_right: BigDecimal,
-    pub normalized_suspension_travel_rear_left: BigDecimal,
-    pub normalized_suspension_travel_rear_right: BigDecimal,
+    pub normalized_suspension_travel_front_left: f32,
+    pub normalized_suspension_travel_front_right: f32,
+    pub normalized_suspension_travel_rear_left: f32,
+    pub normalized_suspension_travel_rear_right: f32,
 
-    pub tire_slip_ratio_front_left: BigDecimal,
-    pub tire_slip_ratio_front_right: BigDecimal,
-    pub tire_slip_ratio_rear_left: BigDecimal,
-    pub tire_slip_ratio_rear_right: BigDecimal,
+    pub tire_slip_ratio_front_left: f32,
+    pub tire_slip_ratio_front_right: f32,
+    pub tire_slip_ratio_rear_left: f32,
+    pub tire_slip_ratio_rear_right: f32,
 
-    pub wheel_rotation_speed_front_left: BigDecimal,
-    pub wheel_rotation_speed_front_right: BigDecimal,
-    pub wheel_rotation_speed_rear_left: BigDecimal,
-    pub wheel_rotation_speed_rear_right: BigDecimal,
+    pub wheel_rotation_speed_front_left: f32,
+    pub wheel_rotation_speed_front_right: f32,
+    pub wheel_rotation_speed_rear_left: f32,
+    pub wheel_rotation_speed_rear_right: f32,
 
     pub wheel_on_rumble_strip_front_left: bool,
     pub wheel_on_rumble_strip_front_right: bool,
     pub wheel_on_rumble_strip_rear_left: bool,
     pub wheel_on_rumble_strip_rear_right: bool,
 
-    pub wheel_in_puddle_front_left: BigDecimal,
-    pub wheel_in_puddle_front_right: BigDecimal,
-    pub wheel_in_puddle_rear_left: BigDecimal,
-    pub wheel_in_puddle_rear_right: BigDecimal,
+    pub wheel_in_puddle_front_left: f32,
+    pub wheel_in_puddle_front_right: f32,
+    pub wheel_in_puddle_rear_left: f32,
+    pub wheel_in_puddle_rear_right: f32,
 
-    pub surface_rumble_front_left: BigDecimal,
-    pub surface_rumble_front_right: BigDecimal,
-    pub surface_rumble_rear_left: BigDecimal,
-    pub surface_rumble_rear_right: BigDecimal,
+    pub surface_rumble_front_left: f32,
+    pub surface_rumble_front_right: f32,
+    pub surface_rumble_rear_left: f32,
+    pub surface_rumble_rear_right: f32,
 
-    pub tire_slip_angle_front_left: BigDecimal,
-    pub tire_slip_angle_front_right: BigDecimal,
-    pub tire_slip_angle_rear_left: BigDecimal,
-    pub tire_slip_angle_rear_right: BigDecimal,
+    pub tire_slip_angle_front_left: f32,
+    pub tire_slip_angle_front_right: f32,
+    pub tire_slip_angle_rear_left: f32,
+    pub tire_slip_angle_rear_right: f32,
 
-    pub tire_combined_slip_front_left: BigDecimal,
-    pub tire_combined_slip_front_right: BigDecimal,
-    pub tire_combined_slip_rear_left: BigDecimal,
-    pub tire_combined_slip_rear_right: BigDecimal,
+    pub tire_combined_slip_front_left: f32,
+    pub tire_combined_slip_front_right: f32,
+    pub tire_combined_slip_rear_left: f32,
+    pub tire_combined_slip_rear_right: f32,
 
-    pub suspension_travel_meters_front_left: BigDecimal,
-    pub suspension_travel_meters_front_right: BigDecimal,
-    pub suspension_travel_meters_rear_left: BigDecimal,
-    pub suspension_travel_meters_rear_right: BigDecimal,
+    pub suspension_travel_meters_front_left: f32,
+    pub suspension_travel_meters_front_right: f32,
+    pub suspension_travel_meters_rear_left: f32,
+    pub suspension_travel_meters_rear_right: f32,
 
-    pub position_x: Option<BigDecimal>,
-    pub position_y: Option<BigDecimal>,
-    pub position_z: Option<BigDecimal>,
+    pub position_x: Option<f32>,
+    pub position_y: Option<f32>,
+    pub position_z: Option<f32>,
 
-    pub speed: Option<BigDecimal>,
-    pub power: Option<BigDecimal>,
-    pub torque: Option<BigDecimal>,
+    pub speed: Option<f32>,
+    pub power: Option<f32>,
+    pub torque: Option<f32>,
 
-    pub tire_temp_front_left: Option<BigDecimal>,
-    pub tire_temp_front_right: Option<BigDecimal>,
-    pub tire_temp_rear_left: Option<BigDecimal>,
-    pub tire_temp_rear_right: Option<BigDecimal>,
+    pub tire_temp_front_left: Option<f32>,
+    pub tire_temp_front_right: Option<f32>,
+    pub tire_temp_rear_left: Option<f32>,
+    pub tire_temp_rear_right: Option<f32>,
 
-    pub boost: Option<BigDecimal>,
-    pub fuel: Option<BigDecimal>,
-    pub distance_traveled: Option<BigDecimal>,
+    pub boost: Option<f32>,
+    pub fuel: Option<f32>,
+    pub distance_traveled: Option<f32>,
 
-    pub best_lap: Option<BigDecimal>,
-    pub last_lap: Option<BigDecimal>,
-    pub current_lap: Option<BigDecimal>,
-    pub current_race_time: Option<BigDecimal>,
+    pub best_lap: Option<f32>,
+    pub last_lap: Option<f32>,
+    pub current_lap: Option<f32>,
+    pub current_race_time: Option<f32>,
     pub lap_number: Option<i32>,
     pub position: Option<i32>,
 
@@ -163,10 +157,10 @@ pub struct ForzaData {
     pub normalized_driving_lane: Option<i32>,
     pub normalized_ai_brake_difference: Option<i32>,
 
-    pub tire_wear_front_left: Option<BigDecimal>,
-    pub tire_wear_front_right: Option<BigDecimal>,
-    pub tire_wear_rear_left: Option<BigDecimal>,
-    pub tire_wear_rear_right: Option<BigDecimal>,
+    pub tire_wear_front_left: Option<f32>,
+    pub tire_wear_front_right: Option<f32>,
+    pub tire_wear_rear_left: Option<f32>,
+    pub tire_wear_rear_right: Option<f32>,
 }
 
 #[non_exhaustive]
